--- Assignment4/README.md
@@ -1,24 +1,0 @@
-Assigment 4 - Jena API
-================
-
-This is the repository for the Jena API lesson. It contains 5 source java files, covering the RDFS, SPARQL and Ontology APIs from Jena. It also contains several RDF files (resoures/) that are referenced and used in the code.
-
-We suggest to use Eclipse IDE, but you can use any other IDE. To start developing, you can go to "File -> Import -> Maven -> Existing Maven Projects" and select the Assigment3 folder. This project uses Maven, and already defines the Jena 3 dependency. This version of the Jena API depends on Java 8, so be sure to configure your environment to use 1.8 level of the compiler (you might need to install the proper JDK).
-
-If you prefer to install Jena on your own, instead of using Maven, check the following instructions:
-
-Installing Jena
-
-* Go to [http://jena.apache.org/download/](http://jena.apache.org/download/)
-
-* Download apache-jena-3.0.0.tar.gz
-
-* In Eclipse, using Java 8: Build Path -> Configure Build Path... -> Add external JARs (Add Jena lib folder)
-
-Once you have configured your environment using either way, you can test it by running the Task00 class.
-
-Check the Jena documentation to start learning about the API:
-
-* Getting started with Apache Jena [https://jena.apache.org/getting_started/](https://jena.apache.org/getting_started/)
-
-* Jena documentation overview (Javadoc)[http://jena.apache.org/documentation/](http://jena.apache.org/documentation/)
