+++ Assignment4/luisgui867-t110217/task07.java
@@ -1,0 +1,82 @@
+import java.io.InputStream;
+
+import org.apache.jena.ontology.Individual;
+import org.apache.jena.ontology.OntClass;
+import org.apache.jena.ontology.OntModel;
+import org.apache.jena.ontology.OntModelSpec;
+import org.apache.jena.ontology.OntResource;
+import org.apache.jena.rdf.model.InfModel;
+import org.apache.jena.rdf.model.ModelFactory;
+import org.apache.jena.rdf.model.RDFNode;
+import org.apache.jena.rdf.model.Statement;
+import org.apache.jena.rdf.model.StmtIterator;
+import org.apache.jena.reasoner.Reasoner;
+import org.apache.jena.reasoner.ReasonerRegistry;
+import org.apache.jena.util.FileManager;
+import org.apache.jena.util.iterator.ExtendedIterator;
+
+/**
+ * Task 07: Querying ontologies (RDFs)
+ * @author elozano
+ * @author isantana
+ *
+ */
+public class Task07
+{
+	public static String ns = "http://somewhere#";
+	
+	public static void main(String args[])
+	{
+		String filename = "resources/example6.rdf";
+		
+		// Create an empty model
+		OntModel model = ModelFactory.createOntologyModel(OntModelSpec.RDFS_MEM_RDFS_INF);
+		
+		// Use the FileManager to find the input file
+		InputStream in = FileManager.get().open(filename);
+	
+		if (in == null)
+			throw new IllegalArgumentException("File: "+filename+" not found");
+	
+		// Read the RDF/XML file
+		model.read(in, null);
+		
+		
+		// ** TASK 7.1: List all individuals of "Person" **
+		OntClass person = model.getOntClass(ns+"Person");
+		ExtendedIterator instances = person.listInstances();
+		
+		while (instances.hasNext())
+		{
+			Individual inst = (Individual) instances.next();
+			System.out.println("Instance of Person: "+inst.getURI());
+		}
+		
+		// ** TASK 7.2: List all subclasses of "Person" **
+		ExtendedIterator subclasses = person.listSubClasses();
+		
+		System.out.println("lista de subclases : ");
+		while (subclasses.hasNext())
+		{
+			OntClass subclass = (OntClass) subclasses.next();
+			System.out.println("Subclass of Person: "+subclass.getURI());
+		}
+		
+
+		// ** TASK 7.3: Make the necessary changes to get as well indirect instances and subclasses. TIP: you need some inference... **
+		System.out.println("\nTASK 7.3 - Make the necessary changes to get as well indirect instances and subclasses\n");
+		Reasoner reasoner = ReasonerRegistry.getOWLReasoner();
+		reasoner = reasoner.bindSchema(model);
+		reasoner.setDerivationLogging(true);
+		InfModel inf = ModelFactory.createInfModel(reasoner, model);
+		StmtIterator i = inf.listStatements(person, null, (RDFNode) null);
+		while(i.hasNext()){
+			Statement s = i.nextStatement();
+			System.out.println("Statement: \n" + s/*.asTriple()*/); //turtle format
+		}
+	
+
+	}
+	
+	
+}
