--- Assignment4/resources/data05.rdf
@@ -1,129 +1,0 @@
-<?xml version="1.0"?>
-
-
-<!DOCTYPE rdf:RDF [
-    <!ENTITY xsd "http://www.w3.org/2001/XMLSchema#" >
-    <!ENTITY rdfs "http://www.w3.org/2000/01/rdf-schema#" >
-    <!ENTITY rdf "http://www.w3.org/1999/02/22-rdf-syntax-ns#" >
-    <!ENTITY example5 "http://data.one.org#" >
-    <!ENTITY foaf "http://xmlns.com/foaf/0.1/" >
-    <!ENTITY vcard-rdf "http://www.w3.org/2001/vcard-rdf/3.0#" >
-]>
-
-
-
-<rdf:RDF xmlns="http://data.five.org#"
-     xml:base="http://data.five.org"
-     xmlns:rdfs="http://www.w3.org/2000/01/rdf-schema#"
-     xmlns:foaf="http://xmlns.com/foaf/0.1/"
-     xmlns:xsd="http://www.w3.org/2001/XMLSchema#"
-     xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-     xmlns:vcard-rdf="http://www.w3.org/2001/vcard-rdf/3.0#">
-  
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Datatypes
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-	 
-    <rdfs:Datatype rdf:about="http://www.w3.org/2001/vcard-rdf/3.0/FN">
-        <rdfs:range rdf:resource="&xsd;string"/>
-    </rdfs:Datatype>
-    
-    <rdfs:Property rdf:about="http://www.w3.org/2001/vcard-rdf/3.0/Family">
-        <rdfs:range rdf:resource="&xsd;string"/>
-    </rdfs:Property>
-
-    <rdfs:Datatype rdf:about="http://www.w3.org/2001/vcard-rdf/3.0/Given">
-        <rdfs:range rdf:resource="&xsd;string"/>
-    </rdfs:Datatype>
-
-    <rdfs:Poperty rdf:about="#sibling">
-        <rdfs:range rdf:resource="#Person"/>
-        <rdfs:domain rdf:resource="#Person"/>
-    </rdfs:Poperty>
-
-    <rdfs:Poperty rdf:about="#parent">
-        <rdfs:range rdf:resource="#Person"/>
-        <rdfs:domain rdf:resource="#Person"/>
-    </rdfs:Poperty>
-
-    <rdfs:Poperty rdf:about="#hasUncleOrAunt">
-        <rdfs:range rdf:resource="#Person"/>
-        <rdfs:domain rdf:resource="#Person"/>
-    </rdfs:Poperty>
-
- 
- 
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Classes
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-    <rdfs:Class rdf:about="#Person"/>
-    
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Individuals
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-
-    <Person rdf:about="#SaraSmith">
-        <vcard-rdf:Given rdf:datatype="&xsd;string">Sara</vcard-rdf:Given>
-        <vcard-rdf:FN rdf:datatype="&xsd;string">Sara Smith</vcard-rdf:FN>
-        <vcard-rdf:Family rdf:datatype="&xsd;string">Smith</vcard-rdf:Family>
-	    <sibling rdf:resource="#JohnSmith"/>
-    </Person>
-    
-
-    <Person rdf:about="#JohnSmith">
-        <vcard-rdf:Given rdf:datatype="&xsd;string">John</vcard-rdf:Given>
-        <vcard-rdf:FN rdf:datatype="&xsd;string">John Smith</vcard-rdf:FN>
-        <vcard-rdf:Family rdf:datatype="&xsd;string">Smith</vcard-rdf:Family>
-	    <sibling rdf:resource="#PeterSmith"/>
-    </Person>
-    
-    <Person rdf:about="#PeterSmith">
-        <vcard-rdf:Given rdf:datatype="&xsd;string">Peter</vcard-rdf:Given>
-        <vcard-rdf:FN rdf:datatype="&xsd;string">Peter Smith</vcard-rdf:FN>
-        <vcard-rdf:Family rdf:datatype="&xsd;string">Smith</vcard-rdf:Family>
-    </Person>
-    
-    <Person rdf:about="#Tim">
-        <vcard-rdf:Given rdf:datatype="&xsd;string">Tim</vcard-rdf:Given>
-        <vcard-rdf:FN rdf:datatype="&xsd;string">Tim Smith</vcard-rdf:FN>
-        <vcard-rdf:Family rdf:datatype="&xsd;string">Smith</vcard-rdf:Family>
-	    <parent rdf:resource="#SaraSmith"/>
-    </Person>
-    
-    <Person rdf:about="#Jack">
-        <vcard-rdf:Given rdf:datatype="&xsd;string">Jack</vcard-rdf:Given>
-        <vcard-rdf:FN rdf:datatype="&xsd;string">Jack Smith</vcard-rdf:FN>
-        <vcard-rdf:Family rdf:datatype="&xsd;string">Smith</vcard-rdf:Family>
-	    <parent rdf:resource="#JohnSmith"/>
-    </Person>
-    
-
-    <Person rdf:about="#Harry">
-        <vcard-rdf:Given rdf:datatype="&xsd;string">Harry</vcard-rdf:Given>
-	    <sibling rdf:resource="#Thomas"/>
-    </Person>
-
-    <Person rdf:about="#Thomas">
-        <vcard-rdf:Given rdf:datatype="&xsd;string">Thomas</vcard-rdf:Given>
-	    <sibling rdf:resource="#Harry"/>
-    </Person>
-</rdf:RDF>
-
-
