--- Assignment4/resources/example.rdf
@@ -1,32 +1,0 @@
-<?xml version="1.0"?>
-<rdf:RDF
-    xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-    xmlns:foaf="http://xmlns.com/foaf/0.1/"
-    xmlns:vcard="http://www.w3.org/2001/vcard-rdf/3.0#">
-  <rdf:Description rdf:about="http://somewhere/JohnSmith">
-    <foaf:knows>
-      <rdf:Description rdf:about="http://somewhere/JaneSmith">
-        <vcard:EMAIL rdf:datatype="http://www.w3.org/2001/XMLSchema#string">jSmith@somewhere.com</vcard:EMAIL>
-        <vcard:FN rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Jane Smith</vcard:FN>
-		<vcard:Given rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Jane</vcard:Given>
-		<vcard:Family rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Smith</vcard:Family>
-		<foaf:email rdf:datatype="http://www.w3.org/2001/XMLSchema#string">jSmith@somewhere.com</foaf:email>
-      </rdf:Description>
-    </foaf:knows>
-    <vcard:FN rdf:datatype="http://www.w3.org/2001/XMLSchema#string">John Smith</vcard:FN>
-	<vcard:Given rdf:datatype="http://www.w3.org/2001/XMLSchema#string">John</vcard:Given>
-	<vcard:Family rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Smith</vcard:Family>
-  </rdf:Description>
-  <rdf:Description rdf:about="http://somewhere/SarahJones">
-    <vcard:FN rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Sarah Jones</vcard:FN>
-    <vcard:Given rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Sarah</vcard:Given>
-	<vcard:Family rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Jones</vcard:Family>
-	<foaf:email rdf:datatype="http://www.w3.org/2001/XMLSchema#string">sJones@somewhere.com</foaf:email>
-  </rdf:Description>
-  <rdf:Description rdf:about="http://somewhere/MattJones">
-    <vcard:FN rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Matt Jones</vcard:FN>
-    <vcard:Given rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Matt</vcard:Given>
-	<vcard:Family rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Jones</vcard:Family>
-	<foaf:knows rdf:resource="http://somewhere/JaneSmith"/>
-  </rdf:Description>
-</rdf:RDF>
