--- Assignment4/resources/example1.rdf
@@ -1,17 +1,0 @@
-<?xml version="1.0"?>
-<rdf:RDF
-    xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-    xmlns:foaf="http://xmlns.com/foaf/0.1/"
-    xmlns:vcard="http://www.w3.org/2001/vcard-rdf/3.0#">
-  <rdf:Description rdf:about="http://somewhere/JohnSmith">
-    <vcard:FN rdf:datatype="http://www.w3.org/2001/XMLSchema#string">John Smith</vcard:FN>
-	<vcard:Given rdf:datatype="http://www.w3.org/2001/XMLSchema#string">John</vcard:Given>
-	<vcard:Family rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Smith</vcard:Family>
-  </rdf:Description>
-  <rdf:Description rdf:about="http://somewhere/SarahJones">
-    <vcard:FN rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Sarah Jones</vcard:FN>
-    <vcard:Given rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Sarah</vcard:Given>
-	<vcard:Family rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Jones</vcard:Family>
-	<foaf:email rdf:datatype="http://www.w3.org/2001/XMLSchema#string">sJones@somewhere.com</foaf:email>
-  </rdf:Description>
-</rdf:RDF>
