--- Assignment4/resources/example2.rdf
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-<rdf:RDF
-    xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-    xmlns:foaf="http://xmlns.com/foaf/0.1/"
-    xmlns:vcard="http://www.w3.org/2001/vcard-rdf/3.0#">
-  <rdf:Description rdf:about="http://somewhere/MattJones">
-    <vcard:FN rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Matt Jones</vcard:FN>
-    <vcard:Given rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Matt</vcard:Given>
-	<vcard:Family rdf:datatype="http://www.w3.org/2001/XMLSchema#string">Jones</vcard:Family>
-	<foaf:knows rdf:resource="http://somewhere/JaneSmith"/>
-  </rdf:Description>
-</rdf:RDF>
