--- Assignment4/resources/example4.rdf
@@ -1,145 +1,0 @@
-<?xml version="1.0"?>
-
-
-<!DOCTYPE rdf:RDF [
-    <!ENTITY xsd "http://www.w3.org/2001/XMLSchema#" >
-    <!ENTITY rdfs "http://www.w3.org/2000/01/rdf-schema#" >
-    <!ENTITY rdf "http://www.w3.org/1999/02/22-rdf-syntax-ns#" >
-    <!ENTITY example4 "http://www.sowhere.com#" >
-    <!ENTITY foaf "http://www.semanticweb.org/ontologies/example4.rdf#http://xmlns.com/foaf/0.1/" >
-    <!ENTITY vcard-rdf "http://www.semanticweb.org/ontologies/example4.rdf#http://www.w3.org/2001/vcard-rdf/3.0/" >
-]>
-
-
-<rdf:RDF xmlns="http://www.sowhere.com#"
-     xml:base="http://www.sowhere.com"
-     xmlns:rdfs="http://www.w3.org/2000/01/rdf-schema#"
-     xmlns:foaf="&example4;http://xmlns.com/foaf/0.1/"
-     xmlns:xsd="http://www.w3.org/2001/XMLSchema#"
-     xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-     xmlns:vcard-rdf="&example4;http://www.w3.org/2001/vcard-rdf/3.0/"
-     xmlns:example4="http://www.semanticweb.org/ontologies/example4.rdf#">
-    
-
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Object Properties
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-
-
-    <!-- http://www.semanticweb.org/ontologies/example4.rdf#http://xmlns.com/foaf/0.1/knows -->
-
-    <rdfs:Property rdf:about="#http://xmlns.com/foaf/0.1/knows"/>
-    
-
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Data properties
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-
-
-    <!-- http://www.semanticweb.org/ontologies/example4.rdf#http://www.w3.org/2001/vcard-rdf/3.0/FN -->
-
-    <rdfs:Datatype rdf:about="#http://www.w3.org/2001/vcard-rdf/3.0/FN">
-        <rdfs:range rdf:resource="&xsd;string"/>
-    </rdfs:Datatype>
-    
-
-
-    <!-- http://www.semanticweb.org/ontologies/example4.rdf#http://www.w3.org/2001/vcard-rdf/3.0/Family -->
-
-    <rdfs:Property rdf:about="#http://www.w3.org/2001/vcard-rdf/3.0/Family">
-        <rdfs:range rdf:resource="&xsd;string"/>
-    </rdfs:Property>
-    
-
-
-    <!-- http://www.semanticweb.org/ontologies/example4.rdf#http://www.w3.org/2001/vcard-rdf/3.0/Given -->
-
-    <rdfs:Datatype rdf:about="#http://www.w3.org/2001/vcard-rdf/3.0/Given">
-        <rdfs:range rdf:resource="&xsd;string"/>
-    </rdfs:Datatype>
-    
-
-
-    <!-- http://www.semanticweb.org/ontologies/example4.rdf#http://xmlns.com/foaf/0.1/email -->
-
-    <rdfs:Datatype rdf:about="#http://xmlns.com/foaf/0.1/email">
-        <rdfs:range rdf:resource="&xsd;string"/>
-    </rdfs:Datatype>
-    
-
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Classes
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-
-
-    <!-- http://www.semanticweb.org/ontologies/example4.rdf#Person -->
-
-    <rdfs:Class rdf:about="#Person"/>
-    
-
-
-    <!-- http://www.semanticweb.org/ontologies/example4.rdf#Researcher -->
-
-    <rdfs:Class rdf:about="#Researcher">
-        <rdfs:subClassOf rdf:resource="#Person"/>
-    </rdfs:Class>
-    
-
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Individuals
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-
-
-    <!-- http://www.semanticweb.org/ontologies/example4.rdf#JaneSmith -->
-
-    <Person rdf:about="#JaneSmith">
-        <vcard-rdf:Given rdf:datatype="&xsd;string">Jane</vcard-rdf:Given>
-        <vcard-rdf:FN rdf:datatype="&xsd;string">Jane Smith</vcard-rdf:FN>
-        <vcard-rdf:Family rdf:datatype="&xsd;string">Smith</vcard-rdf:Family>
-        <foaf:email rdf:datatype="&xsd;string"
-            >jSmith@somewhere.com</foaf:email>
-    </Person>
-    
-
-
-    <!-- http://www.semanticweb.org/ontologies/example4.rdf#JohnSmith -->
-
-    <Person rdf:about="#JohnSmith">
-        <vcard-rdf:Given rdf:datatype="&xsd;string">John</vcard-rdf:Given>
-        <vcard-rdf:FN rdf:datatype="&xsd;string">John Smith</vcard-rdf:FN>
-        <vcard-rdf:Family rdf:datatype="&xsd;string">Smith</vcard-rdf:Family>
-    </Person>
-</rdf:RDF>
-
-
-
-<!-- Generated by the OWL API (version 2.2.1.1138) http://owlapi.sourceforge.net -->
-
