--- Assignment4/resources/example6.rdf
@@ -1,98 +1,0 @@
-<?xml version="1.0"?>
-
-
-<!DOCTYPE rdf:RDF [
-    <!ENTITY xsd "http://www.w3.org/2001/XMLSchema#" >
-    <!ENTITY rdfs "http://www.w3.org/2000/01/rdf-schema#" >
-    <!ENTITY rdf "http://www.w3.org/1999/02/22-rdf-syntax-ns#" >
-    <!ENTITY example6 "http://somewhere#" >
-    <!ENTITY foaf "http://www.semanticweb.org/ontologies/example6.rdf#http://xmlns.com/foaf/0.1/" >
-    <!ENTITY vcard-rdf "http://www.semanticweb.org/ontologies/example6.rdf#http://www.w3.org/2001/vcard-rdf/3.0/" >
-]>
-
-
-<rdf:RDF xmlns="http://somewhere#"
-     xml:base="http://somewhere"
-     xmlns:rdfs="http://www.w3.org/2000/01/rdf-schema#"
-     xmlns:foaf="&example6;http://xmlns.com/foaf/0.1/"
-     xmlns:xsd="http://www.w3.org/2001/XMLSchema#"
-     xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-     xmlns:vcard-rdf="&example6;http://www.w3.org/2001/vcard-rdf/3.0/"
-     xmlns:example6="http://www.semanticweb.org/ontologies/example6.rdf#">
-  
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Datatypes
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-	 
-    <rdfs:Datatype rdf:about="#http://www.w3.org/2001/vcard-rdf/3.0/FN">
-        <rdfs:range rdf:resource="&xsd;string"/>
-    </rdfs:Datatype>
-    
-    <rdfs:Property rdf:about="#http://www.w3.org/2001/vcard-rdf/3.0/Family">
-        <rdfs:range rdf:resource="&xsd;string"/>
-    </rdfs:Property>
-
-    <rdfs:Datatype rdf:about="#http://www.w3.org/2001/vcard-rdf/3.0/Given">
-        <rdfs:range rdf:resource="&xsd;string"/>
-    </rdfs:Datatype>
- 
- 
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Classes
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-    
-    <rdfs:Class rdf:about="#Person"/>
-	
-	<rdfs:Class rdf:about="#Researcher">
-		<rdfs:subClassOf rdf:resource="#Person"/>
-	</rdfs:Class>
-	
-	<rdfs:Class rdf:about="#PhDstudent">
-		<rdfs:subClassOf rdf:resource="#Researcher"/>
-	</rdfs:Class>
-    
-	<rdfs:Class rdf:about="#University"/>
-
-    <!-- 
-    ///////////////////////////////////////////////////////////////////////////////////////
-    //
-    // Individuals
-    //
-    ///////////////////////////////////////////////////////////////////////////////////////
-     -->
-
-
-    <Person rdf:about="#SaraJones">
-        <vcard-rdf:Given rdf:datatype="&xsd;string">Sara</vcard-rdf:Given>
-        <vcard-rdf:FN rdf:datatype="&xsd;string">Sara Jones</vcard-rdf:FN>
-        <vcard-rdf:Family rdf:datatype="&xsd;string">Jones</vcard-rdf:Family>
-    </Person>
-    
-
-    <Person rdf:about="#JohnSmith">
-        <vcard-rdf:Given rdf:datatype="&xsd;string">John</vcard-rdf:Given>
-        <vcard-rdf:FN rdf:datatype="&xsd;string">John Smith</vcard-rdf:FN>
-        <vcard-rdf:Family rdf:datatype="&xsd;string">Smith</vcard-rdf:Family>
-    </Person>
-	
-	<Researcher rdf:about="#JaneSmith">
-        <vcard-rdf:Given rdf:datatype="&xsd;string">Jane</vcard-rdf:Given>
-        <vcard-rdf:FN rdf:datatype="&xsd;string">jane Smith</vcard-rdf:FN>
-        <vcard-rdf:Family rdf:datatype="&xsd;string">Smith</vcard-rdf:Family>
-    </Researcher>
-	
-	<University rdf:about="#UPM"/>
-	
-</rdf:RDF>
-
-
