--- Assignment4/src/main/java/upm/oeg/wsld/jena/App.java
@@ -1,13 +1,0 @@
-package upm.oeg.wsld.jena;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
